+++ code/ext/ext.lib.privy/src/ui/ui.Info/ui.Row.Farcaster.Signer.tsx
